move from trait into abstract cacher
